crawler/requirements.txt
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
